Even more docs/doctests
