+++ src/map/tests.rs
@@ -1,5 +1,5 @@
 use super::*;
-use alloc::{format, vec, vec::Vec};
+use alloc::{collections::BTreeMap, format, vec, vec::Vec};
 
 trait RangeMapExt<K, V> {
     fn to_vec(&self) -> Vec<(Range<K>, V)>;
@@ -12,6 +12,55 @@ where
 {
     fn to_vec(&self) -> Vec<(Range<K>, V)> {
         self.iter().map(|(kr, v)| (kr.clone(), v.clone())).collect()
+    }
+}
+
+// A simple but infeasibly slow and memory-hungry
+// version of `RangeMap` for testing.
+#[derive(Eq, PartialEq, Debug)]
+pub struct StupidU32RangeMap<V> {
+    // Inner B-Tree map. Stores values and their keys
+    // directly rather than as ranges.
+    btm: BTreeMap<u32, V>,
+}
+
+impl<V> StupidU32RangeMap<V>
+where
+    V: Eq + Clone,
+{
+    pub fn new() -> StupidU32RangeMap<V> {
+        StupidU32RangeMap {
+            btm: BTreeMap::new(),
+        }
+    }
+
+    pub fn insert<R: RangeBounds<u32>>(&mut self, range: R, value: V) {
+        let start = match range.start_bound() {
+            Bound::Unbounded => u32::MIN,
+            Bound::Included(&t) => t,
+            Bound::Excluded(&t) => t + 1,
+        };
+        let end = match range.end_bound() {
+            Bound::Unbounded => u32::MAX,
+            Bound::Included(&t) => t,
+            Bound::Excluded(&t) => t - 1,
+        };
+        for k in start..=end {
+            self.btm.insert(k, value.clone());
+        }
+    }
+}
+
+impl<V> From<RangeMap<u32, V>> for StupidU32RangeMap<V>
+where
+    V: Eq + Clone,
+{
+    fn from(range_map: RangeMap<u32, V>) -> Self {
+        let mut stupid = Self::new();
+        for (range, value) in range_map.iter() {
+            stupid.insert(range, value.clone());
+        }
+        stupid
     }
 }
 
@@ -185,7 +234,6 @@ fn replace_at_end_of_existing_range_should_coalesce() {
 #[test]
 // Test every permutation of a bunch of touching and overlapping ranges.
 fn lots_of_interesting_ranges() {
-    use crate::stupid_range_map::StupidU32RangeMap;
     use permutator::Permutation;
 
     let mut ranges_with_values = [
